+++ examples/advanced/src/gen/mocks/createUserFaker.ts
@@ -6,9 +6,10 @@ export function createUserFaker(data?: Partial<User>) {
     ...{
       id: faker.number.int(),
       username: faker.string.alpha(),
+      uuid: faker.string.uuid(),
       firstName: faker.string.alpha(),
       lastName: faker.string.alpha(),
-      email: faker.string.alpha(),
+      email: faker.internet.email(),
       password: faker.string.alpha(),
       phone: faker.string.alpha(),
       userStatus: faker.number.int(),