+++ packages/plugin-faker/src/generators/__snapshots__/petWithRandExp.ts
@@ -3,7 +3,7 @@ import { faker } from "@faker-js/faker";
 
  export function pet(data?: Partial<Pet>) {
     return {
-        ...{ "id": faker.number.int(), "name": faker.string.alpha(), "tag": faker.string.alpha(), "code": faker.helpers.arrayElement<any>([faker.string.alpha(), new RandExp("\\b[1-9]\\b").gen()]), "shipDate": faker.date.anytime(), "shipTime": faker.date.anytime() },
+        ...{ "id": faker.number.int(), "name": faker.string.alpha(), "tag": faker.string.alpha(), "code": new RandExp("\\b[1-9]\\b").gen(), "shipDate": faker.date.anytime(), "shipTime": faker.date.anytime() },
         ...data || {}
     };
 }