+++ packages/plugin-oas/src/SchemaGenerator.ts
@@ -422,13 +422,6 @@ export class SchemaGenerator<
       })
     }
 
-    if (schema.pattern) {
-      baseItems.unshift({
-        keyword: schemaKeywords.matches,
-        args: schema.pattern,
-      })
-    }
-
     if (max !== undefined) {
       baseItems.unshift({ keyword: schemaKeywords.max, args: max })
     }
@@ -793,11 +786,11 @@ export class SchemaGenerator<
           break
         case 'uuid':
           baseItems.unshift({ keyword: schemaKeywords.uuid })
-          break
+          return baseItems
         case 'email':
         case 'idn-email':
           baseItems.unshift({ keyword: schemaKeywords.email })
-          break
+          return baseItems
         case 'uri':
         case 'ipv4':
         case 'ipv6':
@@ -805,7 +798,7 @@ export class SchemaGenerator<
         case 'hostname':
         case 'idn-hostname':
           baseItems.unshift({ keyword: schemaKeywords.url })
-          break
+          return baseItems
         // case 'duration':
         // case 'json-pointer':
         // case 'relative-json-pointer':
@@ -813,6 +806,15 @@ export class SchemaGenerator<
           // formats not yet implemented: ignore.
           break
       }
+    }
+
+    if (schema.pattern) {
+      baseItems.unshift({
+        keyword: schemaKeywords.matches,
+        args: schema.pattern,
+      })
+
+      return baseItems
     }
 
     // type based logic